// src/modules/players/constants/player.enums.ts
export enum PlayerProviderEnum {
  FOOTBALL_DATA = 'football-data',
  API_FOOTBALL = 'api-football',
  MANUAL = 'manual',
}

export enum PlayerPositionEnum {
  GK = 'GK',
  DEF = 'DEF',
  MID = 'MID',
  FW = 'FW',
}

export enum PlayerSyncRunStatusEnum {
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}